+++ SDLS.csproj
@@ -10,11 +10,6 @@
     <LangVersion>latest</LangVersion>
 </PropertyGroup>
 
-<PropertyGroup Condition="'$(DefineConstants.Contains(`SKIES`))'">
-    <!-- Changes the TargetFramework when run in SKIES mode -->
-    <TargetFramework>net48</TargetFramework>
-</PropertyGroup>
-
   <ItemGroup>
     <PackageReference Include="BepInEx.Analyzers" Version="1.*" PrivateAssets="all" />
     <PackageReference Include="BepInEx.Core" Version="5.*" />
@@ -22,12 +17,6 @@
     <PackageReference Include="UnityEngine.Modules" Version="5.5.1" IncludeAssets="compile" />
   </ItemGroup>
 
-  <PropertyGroup>
-  <!-- Supress warning Newtonsoft and other DLL's might generate when they are not compatible with the current TargetFramework -->
-  <NoWarn>$(NoWarn);MSB3258</NoWarn>
-</PropertyGroup>
-
-  
 <ItemGroup Condition="'$(TargetFramework.TrimEnd(`0123456789`))' == 'net'">
     <PackageReference Include="Microsoft.NETFramework.ReferenceAssemblies" Version="1.0.2" PrivateAssets="all" />
     <Reference Include="Sunless.Game">
@@ -36,18 +25,12 @@
     <Reference Include="JsonFx">
         <HintPath>dependencies/JsonFX.dll</HintPath>
     </Reference>
-    <Reference Include="Newtonsoft.Json">
-        <HintPath>dependencies/Newtonsoft.Json.dll</HintPath>
-    </Reference>
 
     <!-- Include the default settings ini file as fallback -->
     <EmbeddedResource Include="SDLS_Config.ini" />
 
     <!-- Include all JSON files in all subdirectories of Sunless Sea -->
-    <EmbeddedResource Include="Sunless Sea/**/*.json" />
-
-    <!-- Include all JSON files in all subdirectories of Sunless Skies -->
-    <EmbeddedResource Include="Sunless Skies/**/*.json" />
+    <EmbeddedResource Include="default/*.json" />
 </ItemGroup>
